feat: restructure product view
